api_bnk v0.0.1 ApiBnK.Web.Router.Helpers

Module with named helpers generated from ApiBnK.Web.Router.

Functions

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.